とりあえず

+++ data/src/client/gameObject/character/character.cpp
@@ -106,7 +106,7 @@ void Character::lookatDir(Vector3f direction)
 
 void Character::weaponThrow(int time)
 {
-    hands[HAND_LEFT]->transform.position += Vector3f(0.0, sin(time), cos(time));
+    hands[HAND_LEFT]->transform.position += Vector3f(0.0, (float)sin(time), (float)cos(time));
 }
 
 void Character::Cancel()
